+++ packages/config/tsconfig/library.json
@@ -19,5 +19,8 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
+  "paths": {
+    "@/*": ["src/*"]
+  },
   "exclude": [".eslintrc.cjs"]
 }
