--- packages/shared/src/__tests__/test.test.ts
@@ -1,4 +1,0 @@
-import { expect, test } from 'vitest';
-test('should work as expected', () => {
-  expect(Math.sqrt(4)).toBe(2);
-});
