+++ packages/shared/src/components/Proxyable/index.ts
@@ -1,0 +1,3 @@
+import Proxyable from './utils/proxyable';
+
+export default Proxyable;
