+++ packages/shared/src/components/index.ts
@@ -1,0 +1,3 @@
+// TODO: Adding this until Proxyable is not the only export
+// eslint-disable-next-line import/prefer-default-export
+export { default as Proxyable } from './Proxyable';
